Refactored: Moved ActionView instrumentation from Rails into integration.
