+++ lib/ddtrace/contrib/action_view/ext.rb
@@ -7,6 +7,10 @@ module Datadog
         ENV_ANALYTICS_ENABLED = 'DD_ACTION_VIEW_ANALYTICS_ENABLED'.freeze
         ENV_ANALYTICS_SAMPLE_RATE = 'DD_ACTION_VIEW_ANALYTICS_SAMPLE_RATE'.freeze
         SERVICE_NAME = 'action_view'.freeze
+        SPAN_RENDER_PARTIAL = 'rails.render_partial'.freeze
+        SPAN_RENDER_TEMPLATE = 'rails.render_template'.freeze
+        TAG_LAYOUT = 'rails.layout'.freeze
+        TAG_TEMPLATE_NAME = 'rails.template_name'.freeze
       end
     end
   end
